Load into SiegeUrls, not into vec
A URLs file *can* be fairly large, so rather than consume unknown,
potentially large amounts of memory, use SiegeUrls which reads each line
lazily in order to consume unknown, but smaller, amounts of memory.

There should be more improvements to come beyond this, as a slow-to-read
file could make the delays longer than intended. For example, being able
to buffer several hundred UrlEntry items ahead of time would be nice, as
well as being able to loop as necessary.