Include base headers
This project allows for a base url, where each url entry only needs to
specify a path, and it gets appended to a url that was specified only
once.

Now, with this, each call can have one or more "base" headers, that is,
headers that will be included in each call. It can be specified with an
env var like so:

    LOADY_HEADERS='x-example-header:foo'

Multiple headers can be specified with a (header) separator:

    LOADY_HEADERS='x-example-header:foo(header)x-example-2:bar'