Allow url list to be divided
By specifying `LOADY_NODE` and `LOADY_NODES`, allows for splitting the
urls list. nodes must be a positive number, and node must be a number
anywhere between 1 and nodes, inclusive. (There was debate on if the
node value should be between 1 and nodes-1, inclusive, but figured that
since pipeline job runners a 1-based, we went with that).