Change defaults and add datetimes
Having a consistent number of threads, regardless of the number of
physical cores, is nice, so the default is 10 threads, but the user can
allow an automatically chosen amount by specifying 0.

Default was for keep-alive connections, but is now close, as having
keep-alive cn introduce interesting effects, like still having
connections up even if the server is not accepting any more connections.
The user should opt-in to that choice.

Lastly, the Datetime of when the job starts and ends is reported in the
final report, and is included periodically throughout the run.